api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
meilisearch==0.28.0
motor==3.3.1
pymongo==4.5.0
requests==2.31.0
httpx==0.25.1
python-multipart==0.0.6
aiohttp==3.8.6
matplotlib==3.8.1
pandas==2.1.2
numpy==1.26.1 